Clear Scene after Unload main window

+++ VisualStudio_2015/GeometryDebugger/GeometryDebugger/Utils/SharedMemory.cs
@@ -42,9 +42,9 @@ namespace GeometryDebugger.Utils
                 {
                     if (m_variables[i].m_B_IsSelected)
                     {
-                        string R = (m_variables[i].m_C_Color.m_i_R / 255).ToString().Replace(",", ".");
-                        string G = (m_variables[i].m_C_Color.m_i_G / 255).ToString().Replace(",", ".");
-                        string B = (m_variables[i].m_C_Color.m_i_B / 255).ToString().Replace(",", ".");
+                        string R = ((float)m_variables[i].m_C_Color.m_i_R / 255).ToString().Replace(",", ".");
+                        string G = ((float)m_variables[i].m_C_Color.m_i_G / 255).ToString().Replace(",", ".");
+                        string B = ((float)m_variables[i].m_C_Color.m_i_B / 255).ToString().Replace(",", ".");
 
                         m_S_message += $"{m_variables[i].m_S_Name}|{m_variables[i].m_S_Type}|{m_variables[i].m_S_Addres}|{R}|{G}|{B}";
                         if (i < m_variables.Count - 1)
